requirements.txt
face_recognition==1.3.0
numpy==1.22.4
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.1.1
requests==2.27.1
streamlit==1.10.0
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
